Remove Sinatra and the web process
The app doesn't use the web process, but I think it was there because
the old, free dyno structure didn't allow for workers without a web
process. Production is now on a Professional worker dyno with no web
dyno, and staging is on a hobby worker dyno with no web dyno. It seems
like bumping up to any of the free tiers removes the web process
restriction and we can now get rid of it.